news_scraper/requirements.txt
python-dateutil==2.6.0
Jinja2==2.10
